Ac(refactor): add account role for improved classification
[parent,child,single] to [root,parent,child,standalone] to better represent
the role of an account.

Every account is either a root or a child.
Among root accounts, an account is either a parent or potentially standalone.
Standalone accounts are specifically root accounts with transactions but no children.

root: category and no parent ac
parent: root ac with children
standalone: root account that is not a parent and have transactions
child: not a root ac and has a parent ac